RavenDB-18880: rename class
