+++ src/Raven.Server/Documents/Sharding/Subscriptions/SubscriptionConnectionsStateOrchestrator.cs
@@ -32,7 +32,7 @@ namespace Raven.Server.Documents.Sharding.Subscriptions;
  * Orchestrator sends the CONFIRM to client
  */
 
-public class SubscriptionConnectionsStateOrchestrator : SubscriptionConnectionsStateBase<OrchestratedSubscriptionConnection, OrchestratorIncludesCommandImpl>
+public class SubscriptionConnectionsStateOrchestrator : AbstractSubscriptionConnectionsState<OrchestratedSubscriptionConnection, OrchestratorIncludesCommandImpl>
 {
     internal readonly ShardedDatabaseContext _databaseContext;
     private Dictionary<string, ShardedSubscriptionWorker> _shardWorkers;
