+++ src/Raven.Server/Documents/Subscriptions/SubscriptionConnectionsState.cs
@@ -22,7 +22,7 @@ using Voron;
 
 namespace Raven.Server.Documents.Subscriptions
 {
-    public class SubscriptionConnectionsState : SubscriptionConnectionsStateBase<SubscriptionConnection, DatabaseIncludesCommandImpl>
+    public class SubscriptionConnectionsState : AbstractSubscriptionConnectionsState<SubscriptionConnection, DatabaseIncludesCommandImpl>
     {
         private readonly SubscriptionStorage _subscriptionStorage;
         public DocumentDatabase DocumentDatabase => _subscriptionStorage._db;
